Centralize deterministic helpers
