+++ warp/_src/codegen.py
@@ -23,6 +23,24 @@ from collections.abc import Callable, Mapping, Sequence
 from typing import Any, ClassVar, get_args, get_origin
 
 import warp.config
+from warp._src.deterministic import (
+    _DET_INTERCEPTABLE_ATOMICS,
+    _DET_UNINTERCEPTED_SIDE_EFFECT_ATOMICS,
+    _add_deterministic_array_store,
+    _det_needs_store_guard,
+    _det_wrap_side_effect_call,
+    _det_wrap_slot_store,
+    _deterministic_adjoint_address_call,
+    _deterministic_call_args,
+    _deterministic_call_name,
+    _deterministic_function_args,
+    _deterministic_has_consumed_atomic_impl,
+    _deterministic_has_propagating_side_effect,
+    _deterministic_kernel_args,
+    _deterministic_kernel_locals,
+    _include_deterministic_call_meta,
+    emit_deterministic_atomic,
+)
 from warp._src.logger import log_debug, log_warning
 from warp._src.types import *
 
@@ -31,50 +49,6 @@ _wp_module_name_ = "warp.codegen"
 # used as a globally accessible copy
 # of current compile options (block_dim) etc
 options = {}
-
-# Atomic builtins that can be intercepted by deterministic mode.
-_DET_INTERCEPTABLE_ATOMICS = frozenset(
-    {
-        "atomic_add",
-        "atomic_sub",
-        "atomic_min",
-        "atomic_max",
-    }
-)
-
-# Atomic builtins that deterministic mode does not intercept, but which still
-# mutate user state and must not double-execute under two-pass counter replay.
-_DET_UNINTERCEPTED_SIDE_EFFECT_ATOMICS = frozenset(
-    {
-        "atomic_xor",
-        "atomic_and",
-        "atomic_or",
-    }
-)
-
-
-def _det_needs_store_guard(adj) -> bool:
-    """Return whether stores in this body must be phase-gated."""
-    if adj.det_meta is None:
-        return False
-    return adj.det_meta.has_counter or adj.det_meta.has_consumed_atomic or adj.det_meta.has_side_effect_store
-
-
-def _det_wrap_slot_store(adj, slot_lvalue: str, value_expr: str) -> str:
-    """Return the forward C++ statement for ``slot_lvalue = value_expr``."""
-    if _det_needs_store_guard(adj):
-        adj.det_meta.needs_context = True
-        return f"WP_DET_SLOT_STORE_IF_ACTIVE(det_ctx, {slot_lvalue}, {value_expr});"
-    return f"{slot_lvalue} = {value_expr};"
-
-
-def _det_wrap_side_effect_call(adj, statement: str) -> str:
-    """Return the forward C++ statement for a non-counter atomic side effect."""
-    if _det_needs_store_guard(adj):
-        adj.det_meta.needs_context = True
-        return f"WP_DET_SIDE_EFFECT_IF_ACTIVE(det_ctx, {statement});"
-    return statement
-
 
 def _escape_line_directive_filename(filename: str) -> str:
     """Return ``filename`` escaped for the quoted filename field of a C/CUDA ``#line`` directive."""
@@ -96,165 +70,6 @@ def _escape_line_directive_filename(filename: str) -> str:
             escaped.append(c)
 
     return "".join(escaped)
-
-
-def _deterministic_call_name(node):
-    if isinstance(node, ast.Attribute):
-        return node.attr
-    if isinstance(node, ast.Name):
-        return node.id
-    return None
-
-
-def _deterministic_contains_atomic_call(node):
-    for child in ast.walk(node):
-        if isinstance(child, ast.Call) and _deterministic_call_name(child.func) in _DET_INTERCEPTABLE_ATOMICS:
-            return True
-    return False
-
-
-def _deterministic_contains_subscript_target(node):
-    if isinstance(node, ast.Subscript):
-        return True
-    if isinstance(node, (ast.Tuple, ast.List)):
-        return any(_deterministic_contains_subscript_target(element) for element in node.elts)
-    return False
-
-
-def _deterministic_has_propagating_side_effect(tree):
-    """Return whether ``tree`` contains an array store or unintercepted bitwise atomic."""
-    for node in ast.walk(tree):
-        if isinstance(node, ast.Assign):
-            if any(_deterministic_contains_subscript_target(target) for target in node.targets):
-                return True
-        elif isinstance(node, (ast.AnnAssign, ast.AugAssign)):
-            if _deterministic_contains_subscript_target(node.target):
-                return True
-        elif isinstance(node, ast.Call):
-            if _deterministic_call_name(node.func) in _DET_UNINTERCEPTED_SIDE_EFFECT_ATOMICS:
-                return True
-
-    return False
-
-
-def _deterministic_has_consumed_atomic(tree):
-    """Return whether codegen will treat any atomic return value as consumed."""
-    return _deterministic_has_consumed_atomic_impl(tree, adj=None, seen=None)
-
-
-def _deterministic_collect_target_names(target, names):
-    if isinstance(target, ast.Name):
-        names.add(target.id)
-    elif isinstance(target, (ast.Tuple, ast.List)):
-        for element in target.elts:
-            _deterministic_collect_target_names(element, names)
-
-
-def _deterministic_local_names(tree):
-    names = set()
-
-    for node in ast.walk(tree):
-        if isinstance(node, (ast.FunctionDef, ast.AsyncFunctionDef, ast.Lambda)):
-            args = node.args
-            for arg in itertools.chain(args.posonlyargs, args.args, args.kwonlyargs):
-                names.add(arg.arg)
-            if args.vararg is not None:
-                names.add(args.vararg.arg)
-            if args.kwarg is not None:
-                names.add(args.kwarg.arg)
-        elif isinstance(node, ast.Assign):
-            for target in node.targets:
-                _deterministic_collect_target_names(target, names)
-        elif isinstance(node, ast.AnnAssign):
-            _deterministic_collect_target_names(node.target, names)
-        elif isinstance(node, ast.NamedExpr):
-            _deterministic_collect_target_names(node.target, names)
-        elif isinstance(node, (ast.For, ast.AsyncFor)):
-            _deterministic_collect_target_names(node.target, names)
-        elif isinstance(node, ast.With):
-            for item in node.items:
-                if item.optional_vars is not None:
-                    _deterministic_collect_target_names(item.optional_vars, names)
-        elif isinstance(node, ast.ExceptHandler) and isinstance(node.name, str):
-            names.add(node.name)
-
-    return names
-
-
-def _deterministic_call_path(node):
-    path = []
-
-    while isinstance(node, ast.Attribute):
-        path.append(node.attr)
-        node = node.value
-
-    if isinstance(node, ast.Name):
-        path.append(node.id)
-        path.reverse()
-        return path
-
-    return None
-
-
-def _deterministic_resolve_static_call(adj, node, local_names):
-    if adj is None:
-        return None
-
-    path = _deterministic_call_path(node)
-    if not path or path[0] in local_names:
-        return None
-
-    obj = adj.resolve_external_reference(path[0])
-
-    if obj is None:
-        obj = getattr(warp, path[0], None)
-
-    if obj is None:
-        builtins_obj = __builtins__
-        if isinstance(builtins_obj, dict):
-            obj = builtins_obj.get(path[0])
-        else:
-            obj = getattr(builtins_obj, path[0], None)
-
-    if obj is None:
-        return None
-
-    for attr in path[1:]:
-        if not hasattr(obj, attr):
-            return None
-        obj = getattr(obj, attr)
-
-    return obj
-
-
-def _deterministic_has_consumed_atomic_impl(tree, adj, seen):
-    """Return whether this AST or any statically called Warp function consumes an atomic return."""
-    if seen is None:
-        seen = set()
-
-    local_names = _deterministic_local_names(tree)
-
-    stack = [(tree, None)]
-    while stack:
-        node, parent = stack.pop()
-
-        if isinstance(node, ast.Call):
-            if _deterministic_call_name(node.func) in _DET_INTERCEPTABLE_ATOMICS:
-                if not (isinstance(parent, ast.Expr) and parent.value is node):
-                    return True
-            else:
-                func = _deterministic_resolve_static_call(adj, node.func, local_names)
-                if isinstance(func, warp._src.context.Function) and not func.is_builtin():
-                    func_id = id(func)
-                    if func_id not in seen:
-                        seen.add(func_id)
-                        if _deterministic_has_consumed_atomic_impl(func.adj.tree, func.adj, seen):
-                            return True
-
-        for child in ast.iter_child_nodes(node):
-            stack.append((child, node))
-
-    return False
 
 
 def get_node_name_safe(node):
@@ -2101,7 +1916,7 @@ class Adjoint:
         # Deterministic mode: intercept atomic builtins and emit scatter or
         # two-pass code instead of the normal atomic call.
         if adj.det_meta is not None and func.is_builtin() and func.key in _DET_INTERCEPTABLE_ATOMICS:
-            det_output = adj._emit_deterministic_atomic(func, bound_args, return_type, output, output_list)
+            det_output = emit_deterministic_atomic(adj, func, bound_args, return_type, output, output_list)
             if det_output is not None:
                 return det_output
 
@@ -2240,213 +2055,6 @@ class Adjoint:
     def add_builtin_call(adj, func_name, args, min_outputs=None):
         func = warp._src.context.builtin_functions[func_name]
         return adj.add_call(func, args, {}, {}, min_outputs=min_outputs)
-
-    def _emit_deterministic_atomic(adj, func, bound_args, return_type, output, output_list):
-        """Emit deterministic scatter or two-pass code for an atomic builtin.
-
-        Returns the output Var if the atomic was handled, or None only for
-        atomic operations whose ordinary implementation is already
-        deterministic.
-        """
-        from warp._src.deterministic import (  # noqa: PLC0415
-            REDUCE_OP_ADD,
-            REDUCE_OP_MAX,
-            REDUCE_OP_MIN,
-            get_or_create_counter_target,
-            get_or_create_scatter_target,
-        )
-
-        args_list = list(bound_args.values())
-        arr_var = args_list[0]  # the target array, possibly a view such as arr[i]
-        view_prefix_vars = []
-
-        while getattr(arr_var, "_det_view_parent", None) is not None:
-            view_prefix_vars = list(arr_var._det_view_indices) + view_prefix_vars
-            arr_var = arr_var._det_view_parent
-
-        try:
-            target_info = _deterministic_target_info(arr_var)
-            if target_info is None:
-                raise WarpCodegenError(f"Deterministic mode could not resolve the target array for {func.key}.")
-
-            target_root_label, target_attr_path, arr_type, target_is_adjoint = target_info
-            if not is_array(arr_type):
-                raise WarpCodegenError(
-                    f"Deterministic mode expected {func.key} to target an array, got {type_repr(arr_type)}."
-                )
-
-            value_dtype = arr_type.dtype
-            scalar_dtype = value_dtype
-            if hasattr(scalar_dtype, "_wp_scalar_type_"):
-                scalar_dtype = scalar_dtype._wp_scalar_type_
-
-            if not any(arg.label == target_root_label for arg in adj.args):
-                raise WarpCodegenError(
-                    f"Deterministic mode could not map atomic target '{target_root_label}' "
-                    f"for {func.key} to a kernel argument."
-                )
-        except WarpCodegenError:
-            raise
-        except (AttributeError, KeyError, TypeError, ValueError) as e:
-            raise WarpCodegenError(f"Deterministic mode could not lower {func.key}: {e}") from e
-
-        # Determine if the return value is actually consumed by the caller.
-        # When called from emit_AugAssign (arr[i] += val) or a bare expression,
-        # the return is discarded and the atomic can be handled by scatter/reduce.
-        # Any other expression context consumes the return value.
-        return_is_discarded = getattr(adj, "_det_atomic_return_discarded", False)
-        return_is_consumed = not return_is_discarded
-
-        value_ctype = Var.dtype_to_ctype(value_dtype)
-        scalar_ctype = Var.dtype_to_ctype(scalar_dtype)
-
-        zero_expr = _cinit_expr(return_type)
-
-        # Map from builtin name to reduction op
-        op_map = {
-            "atomic_add": REDUCE_OP_ADD,
-            "atomic_sub": REDUCE_OP_ADD,  # value is negated before scatter
-            "atomic_min": REDUCE_OP_MIN,
-            "atomic_max": REDUCE_OP_MAX,
-        }
-        reduce_op = op_map.get(func.key, REDUCE_OP_ADD)
-
-        # Classify whether this atomic needs deterministic slot assignment or
-        # can be deferred into the post-kernel scatter/reduce pass.
-        # All deterministic codegen is wrapped in #ifdef __CUDA_ARCH__ so that
-        # CPU compilation falls through to normal atomic calls (CPU execution
-        # is already sequential/deterministic).
-        # We return None for the #else branch, telling add_call to emit the
-        # normal atomic call below. However, since we can only return once,
-        # we handle both CUDA and CPU paths here: emit the CUDA path as
-        # raw C++ with #ifdef, then return output to skip the normal codegen.
-
-        # Build the CPU fallback: a normal atomic call string.
-        # This is used in the #else branch for CPU compilation.
-        loaded_args = [adj.load(a) for a in args_list]
-        cpu_args_str = ", ".join(a.emit() for a in loaded_args)
-        if output is not None:
-            cpu_call = f"var_{output} = wp::{func.native_func}({cpu_args_str});"
-        else:
-            cpu_call = f"wp::{func.native_func}({cpu_args_str});"
-
-        # Integer accumulation atomics are deterministic in a single pass, but
-        # in two-pass counter kernels they are side effects and must not run
-        # during phase 0.  The pre-scan handles atomics that appear before the
-        # counter atomic in source order.
-        if scalar_dtype not in float_types and not return_is_consumed:
-            if adj.det_meta.has_consumed_atomic:
-                adj.add_forward(f"WP_DET_SIDE_EFFECT_IF_ACTIVE(det_ctx, {cpu_call});")
-                return output
-            return None  # fall through to normal codegen
-
-        if return_is_consumed:
-            # Consumed-return counter: assign slots by replaying after a prefix sum.
-            if func.key != "atomic_add":
-                raise WarpCodegenError(
-                    "Deterministic mode currently supports consumed-return counter atomics only for atomic_add, "
-                    f"got {func.key}."
-                )
-
-            if scalar_dtype != int32:
-                raise WarpCodegenError(
-                    "Deterministic mode currently supports consumed-return counter atomics only for int32 "
-                    "counter arrays."
-                )
-
-            try:
-                target = get_or_create_counter_target(
-                    adj.det_registry,
-                    adj.det_meta,
-                    target_root_label,
-                    scalar_ctype,
-                    attr_path=target_attr_path,
-                    adjoint=target_is_adjoint,
-                    use_forward=not adj.custom_reverse_mode and not target_is_adjoint,
-                    use_backward=adj.custom_reverse_mode or target_is_adjoint,
-                )
-            except ValueError as e:
-                raise WarpCodegenError(str(e)) from e
-            helper_name = target.helper_name
-
-            val_loaded = loaded_args[-1]  # already loaded above
-            loaded_prefix = [adj.load(var) for var in view_prefix_vars]
-            idx_loaded_list = loaded_prefix + list(loaded_args[1:-1])
-            target_expr = _deterministic_array_expr(adj, target_root_label, target_attr_path)
-            if target_expr is None:
-                raise WarpCodegenError(
-                    f"Deterministic mode could not build a generated target expression for {func.key} "
-                    f"on '{target_root_label}'."
-                )
-            flat_idx_expr = _deterministic_flat_index_expr(
-                adj, target_expr, idx_loaded_list, func.key, target.value_ctype
-            )
-
-            adj.add_forward(
-                f"WP_DET_COUNTER_OR_FALLBACK(var_{output}, det_ctx, {helper_name}, {flat_idx_expr}, "
-                f"{val_loaded.emit()}, {cpu_call});",
-                replay="// deterministic counter replay (skipped)",
-            )
-            return output
-
-        # Return value unused: record the update and reduce it after the kernel.
-        try:
-            target = get_or_create_scatter_target(
-                adj.det_registry,
-                adj.det_meta,
-                target_root_label,
-                value_dtype,
-                value_ctype,
-                scalar_dtype,
-                reduce_op,
-                attr_path=target_attr_path,
-                adjoint=target_is_adjoint,
-                use_forward=not adj.custom_reverse_mode and not target_is_adjoint,
-                use_backward=adj.custom_reverse_mode or target_is_adjoint,
-            )
-        except ValueError as e:
-            raise WarpCodegenError(str(e)) from e
-        helper_name = target.helper_name
-
-        val_loaded = loaded_args[-1]
-        loaded_prefix = [adj.load(var) for var in view_prefix_vars]
-        idx_loaded_list = loaded_prefix + list(loaded_args[1:-1])
-
-        target_expr = _deterministic_array_expr(adj, target_root_label, target_attr_path)
-        if target_expr is None:
-            raise WarpCodegenError(
-                f"Deterministic mode could not build a generated target expression for {func.key} "
-                f"on '{target_root_label}'."
-            )
-
-        flat_idx_expr = _deterministic_flat_index_expr(adj, target_expr, idx_loaded_list, func.key, target.value_ctype)
-
-        val_expr = val_loaded.emit()
-        if func.key == "atomic_sub":
-            val_expr = f"(-{val_expr})"
-
-        adj.add_forward(
-            f"WP_DET_SCATTER_OR_FALLBACK(det_ctx, {helper_name}, {flat_idx_expr}, {val_expr}, {cpu_call});",
-            replay="// deterministic scatter replay (skipped)",
-        )
-        if output is not None:
-            adj.add_forward(f"var_{output} = {zero_expr};")
-
-        if target_is_adjoint:
-            return output
-
-        target_adj_expr = _deterministic_array_expr(adj, target_root_label, target_attr_path, adjoint=True)
-        if target_adj_expr is None:
-            return output
-
-        original_val_arg = args_list[-1]
-        fwd_parts = [target_expr] + [var.emit() for var in idx_loaded_list] + [val_loaded.emit()]
-        adj_ret_str = f"adj_{output}" if output is not None else zero_expr
-        adj_parts = (
-            [target_adj_expr] + [var.emit_adj() for var in idx_loaded_list] + [original_val_arg.emit_adj(), adj_ret_str]
-        )
-        adj.add_reverse(f"wp::adj_{func.native_func}({', '.join(fwd_parts + adj_parts)});")
-        return output
 
     def add_grad_call(adj, func, args, kwargs):
         """Generate code for calling the gradient of a function via warp.grad().
@@ -6078,423 +5686,6 @@ def codegen_snippet(adj, name, snippet, adj_snippet, replay_snippet, forward_onl
         )
 
     return s
-
-
-def _deterministic_function_args(adj):
-    """Return hidden deterministic parameters for generated ``@wp.func`` calls."""
-    if adj.det_meta is None or not adj.det_meta.needs_deterministic:
-        return []
-
-    from warp._src.deterministic import counter_cpp_type, scatter_cpp_type  # noqa: PLC0415
-
-    det_args = ["wp::det_ctx det_ctx"]
-    for target in adj.det_meta.counter_targets:
-        det_args.append(f"{counter_cpp_type(target)} {target.helper_name}")
-    for target in adj.det_meta.scatter_targets:
-        det_args.append(f"{scatter_cpp_type(target)} {target.helper_name}")
-    return det_args
-
-
-def _deterministic_kernel_args(adj):
-    """Return raw deterministic launch parameters for generated CUDA kernels."""
-    if adj.det_meta is None or not adj.det_meta.needs_deterministic:
-        return []
-
-    from warp._src.deterministic import kernel_raw_counter_param_names, kernel_raw_scatter_param_names  # noqa: PLC0415
-
-    det_args = [
-        "int _wp_det_phase",
-        "int _wp_det_debug",
-        "int* _wp_det_overflow",
-    ]
-    for target in adj.det_meta.counter_targets:
-        names = kernel_raw_counter_param_names(target)
-        det_args.append(f"wp::det_counter_buf_t {names['buf']}")
-    for target in adj.det_meta.scatter_targets:
-        names = kernel_raw_scatter_param_names(target)
-        det_args.append(f"wp::det_scatter_buf_t<{target.value_ctype}> {names['buf']}")
-    return det_args
-
-
-def _deterministic_kernel_locals(adj, device, *, use_launch_buffers=True):
-    """Declare ``det_ctx`` plus helper aliases/stand-ins inside generated kernels."""
-    if adj.det_meta is None or not adj.det_meta.needs_deterministic:
-        return ""
-
-    from warp._src.deterministic import kernel_raw_counter_param_names, kernel_raw_scatter_param_names  # noqa: PLC0415
-
-    if device == "cuda" and use_launch_buffers:
-        decls = []
-        n_counter_targets = len(adj.det_meta.counter_targets)
-        if n_counter_targets > 0:
-            counter_buf_names = [
-                kernel_raw_counter_param_names(target)["buf"] for target in adj.det_meta.counter_targets
-            ]
-            ptr_inits = ", ".join(f"{name}.target_ptr" for name in counter_buf_names)
-            size_inits = ", ".join(f"{name}.target_size" for name in counter_buf_names)
-            decls.append(f"uint64_t _wp_det_counter_target_ptrs[{n_counter_targets}] = {{{ptr_inits}}};")
-            decls.append(f"int _wp_det_counter_target_sizes[{n_counter_targets}] = {{{size_inits}}};")
-            decls.append(
-                "wp::det_ctx det_ctx{_wp_det_phase, _wp_det_debug, _idx, _wp_det_overflow, "
-                f"_wp_det_counter_target_ptrs, _wp_det_counter_target_sizes, {n_counter_targets}}};"
-            )
-        else:
-            decls.append(
-                "wp::det_ctx det_ctx{_wp_det_phase, _wp_det_debug, _idx, _wp_det_overflow, nullptr, nullptr, 0};"
-            )
-        for target in adj.det_meta.counter_targets:
-            names = kernel_raw_counter_param_names(target)
-            decls.append(f"auto& {target.helper_name} = {names['buf']};")
-        for target in adj.det_meta.scatter_targets:
-            names = kernel_raw_scatter_param_names(target)
-            decls.append(f"auto& {target.helper_name} = {names['buf']};")
-    else:
-        decls = [
-            "wp::det_ctx det_ctx{1, 0, 0, nullptr, nullptr, nullptr, 0};",
-        ]
-        for target in adj.det_meta.counter_targets:
-            decls.append(
-                f"wp::det_counter_buf_t {target.helper_name}{{nullptr, nullptr, nullptr, nullptr, nullptr, nullptr, 0, 0, 0, 0}};"
-            )
-        for target in adj.det_meta.scatter_targets:
-            decls.append(
-                f"wp::det_scatter_buf_t<{target.value_ctype}> {target.helper_name}{{nullptr, nullptr, nullptr, 0}};"
-            )
-    return "".join(f"    {line}\n" for line in decls)
-
-
-def _deterministic_reference_origin(var):
-    """Return the root argument and field path for a tracked array reference."""
-    root_label = getattr(var, "_det_ref_root_label", None)
-    if root_label is None:
-        return None
-    return root_label, tuple(getattr(var, "_det_ref_attr_path", ()))
-
-
-def _deterministic_array_expr(adj, root_label, attr_path, adjoint=False):
-    """Build a generated C++ expression for a direct or struct-field array."""
-    root = None
-    for arg in adj.args:
-        if arg.label == root_label:
-            root = arg
-            break
-
-    if root is None:
-        return None
-
-    expr = root.emit_adj() if adjoint else root.emit()
-    for attr in attr_path:
-        expr = f"{expr}.{attr}"
-    return expr
-
-
-def _deterministic_target_info(var):
-    """Return ``(root_arg, attr_path, array_type, adjoint)`` for a deterministic target."""
-    var_type = getattr(var, "type", None)
-    if var_type is None:
-        return None
-
-    if getattr(var, "_det_adjoint_target", False):
-        root_label = getattr(var, "_det_adjoint_root_label", None)
-        if root_label is None:
-            return None
-        attr_path = tuple(getattr(var, "_det_ref_attr_path", ()))
-        array_type = getattr(var, "_det_ref_array_type", strip_reference(var_type))
-        return root_label, attr_path, array_type, True
-
-    origin = _deterministic_reference_origin(var)
-    if origin is not None:
-        root_label, attr_path = origin
-        array_type = getattr(var, "_det_ref_array_type", strip_reference(var_type))
-        return root_label, attr_path, array_type, False
-
-    if is_reference(var_type):
-        return None
-
-    return var.label, (), var_type, False
-
-
-def _add_deterministic_array_store(adj, target, indices, rhs):
-    """Emit ``array_store`` guarded so counter phase 0 has no side effects."""
-    store_args_raw = (target, *indices, rhs)
-    store_func = adj.resolve_func(
-        warp._src.context.builtin_functions["array_store"],
-        tuple(get_arg_type(x) for x in store_args_raw),
-        {},
-        min_outputs=None,
-    )
-    bound_args = store_func.signature.bind(*store_args_raw)
-    bound_arg_types = {k: get_arg_type(v) for k, v in bound_args.arguments.items()}
-    bound_arg_values = {k: get_arg_value(v) for k, v in bound_args.arguments.items()}
-    return_type = store_func.value_func(
-        {k: strip_reference(v) for k, v in bound_arg_types.items()},
-        bound_arg_values,
-    )
-
-    if store_func.dispatch_func is not None:
-        func_args, template_args = store_func.dispatch_func(store_func.input_types, return_type, bound_args)
-    else:
-        func_args = tuple(bound_args.arguments.values())
-        template_args = ()
-
-    func_args = tuple(adj.register_var(x) for x in func_args)
-    func_name = compute_type_str(store_func.native_func, template_args)
-    use_initializer_list = store_func.initializer_list_func(bound_args, return_type)
-
-    # Reuse normal array_store dispatch/loading so references, views, and
-    # adjoint code match the non-deterministic path exactly.
-    fwd_args = []
-    for func_arg in func_args:
-        loaded_arg = func_arg
-        if not isinstance(func_arg, (Reference, warp._src.context.Function)):
-            loaded_arg = adj.load(func_arg)
-        fwd_args.append(strip_reference(loaded_arg))
-
-    store_args = ", ".join(f"var_{x}" for x in fwd_args)
-    adj.add_forward(f"WP_DET_STORE_IF_ACTIVE(det_ctx, {store_args});", skip_replay=True)
-
-    if store_func.is_differentiable:
-        adj_args = tuple(strip_reference(x) for x in func_args)
-        arg_str = adj.format_reverse_call_args(
-            fwd_args,
-            adj_args,
-            [],
-            use_initializer_list,
-            has_output_args=False,
-        )
-        if arg_str is not None:
-            adj.add_reverse(f"{store_func.namespace}adj_{func_name}({arg_str});")
-
-
-def _deterministic_adjoint_address_call(adj, fwd_args, output_list, fallback_call):
-    """Return a deterministic reverse ``address`` call when it scatters into an array adjoint.
-
-    Reverse-mode array reads accumulate into ``adj_arr[index]`` (or
-    ``arr.grad[index]``) with native floating-point atomics.  Under deterministic
-    mode, route those updates through the same scatter/sort/reduce helper used
-    by forward scatter atomics so tape backward passes get a fixed reduction
-    order.
-    """
-    if not fwd_args or not output_list:
-        return fallback_call
-
-    from warp._src.deterministic import (  # noqa: PLC0415
-        REDUCE_OP_ADD,
-        get_or_create_scatter_target,
-        warp_scalar_type_to_id,
-    )
-
-    arr_var = fwd_args[0]
-    view_prefix_vars = []
-    while getattr(arr_var, "_det_view_parent", None) is not None:
-        view_prefix_vars = list(arr_var._det_view_indices) + view_prefix_vars
-        arr_var = arr_var._det_view_parent
-
-    try:
-        target_info = _deterministic_target_info(arr_var)
-        if target_info is None:
-            return fallback_call
-
-        target_root_label, target_attr_path, arr_type, _target_is_adjoint = target_info
-        if not is_array(arr_type):
-            return fallback_call
-
-        value_dtype = arr_type.dtype
-        scalar_dtype = value_dtype
-        if hasattr(scalar_dtype, "_wp_scalar_type_"):
-            scalar_dtype = scalar_dtype._wp_scalar_type_
-        if scalar_dtype not in float_types:
-            return fallback_call
-
-        if not any(arg.label == target_root_label for arg in adj.args):
-            return fallback_call
-
-        value_ctype = Var.dtype_to_ctype(value_dtype)
-        target = get_or_create_scatter_target(
-            adj.det_registry,
-            adj.det_meta,
-            target_root_label,
-            value_dtype,
-            value_ctype,
-            scalar_dtype,
-            REDUCE_OP_ADD,
-            attr_path=target_attr_path,
-            adjoint=True,
-            use_forward=False,
-            use_backward=True,
-        )
-        # Validate that the scalar value has a deterministic reducer.  The
-        # result is used by launch-time post-reduce, but checking here gives a
-        # source-adjacent error if an unsupported gradient type appears.
-        warp_scalar_type_to_id(scalar_dtype)
-    except (AttributeError, KeyError, TypeError, ValueError) as e:
-        raise WarpCodegenError(f"Deterministic mode could not lower adjoint address accumulation: {e}") from e
-
-    loaded_prefix = [adj.load(var) for var in view_prefix_vars]
-    idx_loaded_list = loaded_prefix + list(fwd_args[1:])
-    target_expr = _deterministic_array_expr(adj, target_root_label, target_attr_path)
-    if target_expr is None:
-        return fallback_call
-
-    flat_idx_expr = _deterministic_flat_index_expr(adj, target_expr, idx_loaded_list, "address", target.value_ctype)
-    adj_output = output_list[0].emit_adj()
-    return f"WP_DET_SCATTER_OR_FALLBACK(det_ctx, {target.helper_name}, {flat_idx_expr}, {adj_output}, {fallback_call});"
-
-
-def _deterministic_flat_index_expr(adj, target_expr, idx_loaded_list, func_key, value_ctype):
-    """Element offset from ``ptr`` via ``strides``, so non-contiguous views work."""
-    del adj
-
-    ndim = len(idx_loaded_list)
-    if ndim < 1 or ndim > 4:
-        raise WarpCodegenError(
-            f"Deterministic mode currently supports arrays up to 4 dimensions, got {ndim}D indexing for "
-            f"{func_key} on {target_expr}."
-        )
-
-    terms = " + ".join(f"var_{idx_loaded_list[k]} * {target_expr}.strides[{k}]" for k in range(ndim))
-    return f"(({terms}) / static_cast<int>(sizeof({value_ctype})))"
-
-
-def _deterministic_bound_target(bound_var, extra_attr_path=()):
-    """Map a formal target to the actual bound argument at a call site."""
-    bound_type = getattr(bound_var, "type", None)
-    if bound_type is None:
-        return None
-
-    extra_attr_path = tuple(extra_attr_path)
-    origin = _deterministic_reference_origin(bound_var)
-    if origin is not None:
-        root_label, attr_path = origin
-        return root_label, attr_path + extra_attr_path
-
-    if is_reference(bound_type):
-        return None
-
-    return bound_var.label, extra_attr_path
-
-
-def _deterministic_map_target(target, bound_args):
-    """Remap callee metadata targets into the caller's argument namespace."""
-    attr_path = tuple(getattr(target, "attr_path", ()))
-    actual = bound_args.get(target.array_var_label)
-    if actual is None:
-        return target.array_var_label, attr_path
-
-    mapped = _deterministic_bound_target(strip_reference(actual), attr_path)
-    if mapped is None:
-        return target.array_var_label, attr_path
-
-    return mapped
-
-
-def _deterministic_find_target(targets, array_var_label, attr_path, *, adjoint=False):
-    attr_path = tuple(attr_path)
-    for target in targets:
-        if (
-            target.array_var_label == array_var_label
-            and tuple(getattr(target, "attr_path", ())) == attr_path
-            and bool(getattr(target, "adjoint", False)) == bool(adjoint)
-        ):
-            return target
-    return None
-
-
-def _include_deterministic_call_meta(adj, meta, bound_args):
-    """Merge deterministic requirements from a called ``@wp.func`` into ``adj``."""
-    if adj.det_meta is None or adj.det_registry is None or meta is None or not meta.needs_deterministic:
-        return
-
-    from warp._src.deterministic import get_or_create_counter_target, get_or_create_scatter_target  # noqa: PLC0415
-
-    adj.det_meta.needs_context |= meta.needs_context
-
-    for target, count in meta.scatter_records_per_thread.items():
-        mapped_label, mapped_attr_path = _deterministic_map_target(target, bound_args)
-        mapped_target = get_or_create_scatter_target(
-            adj.det_registry,
-            adj.det_meta,
-            mapped_label,
-            target.value_dtype,
-            target.value_ctype,
-            target.scalar_dtype,
-            target.reduce_op,
-            attr_path=mapped_attr_path,
-            adjoint=getattr(target, "adjoint", False),
-            use_forward=getattr(target, "use_forward", True),
-            use_backward=getattr(target, "use_backward", False),
-        )
-        adj.det_meta.scatter_records_per_thread[mapped_target] += count - 1
-
-    for target, count in meta.counter_records_per_thread.items():
-        mapped_label, mapped_attr_path = _deterministic_map_target(target, bound_args)
-        mapped_target = get_or_create_counter_target(
-            adj.det_registry,
-            adj.det_meta,
-            mapped_label,
-            target.value_ctype,
-            attr_path=mapped_attr_path,
-            adjoint=getattr(target, "adjoint", False),
-            use_forward=getattr(target, "use_forward", True),
-            use_backward=getattr(target, "use_backward", False),
-        )
-        adj.det_meta.counter_records_per_thread[mapped_target] += count - 1
-
-
-def _deterministic_call_args(adj, meta, bound_args):
-    """Return helper arguments to pass from a caller into a deterministic ``@wp.func``."""
-    if adj.det_meta is None or meta is None or not meta.needs_deterministic:
-        return []
-
-    from warp._src.deterministic import get_or_create_counter_target, get_or_create_scatter_target  # noqa: PLC0415
-
-    det_args = ["det_ctx"]
-    for target in meta.counter_targets:
-        mapped_label, mapped_attr_path = _deterministic_map_target(target, bound_args)
-        mapped_target = _deterministic_find_target(
-            adj.det_meta.counter_targets, mapped_label, mapped_attr_path, adjoint=getattr(target, "adjoint", False)
-        )
-        if mapped_target is None and adj.det_registry is not None:
-            mapped_target = get_or_create_counter_target(
-                adj.det_registry,
-                adj.det_meta,
-                mapped_label,
-                target.value_ctype,
-                attr_path=mapped_attr_path,
-                adjoint=getattr(target, "adjoint", False),
-                use_forward=getattr(target, "use_forward", True),
-                use_backward=getattr(target, "use_backward", False),
-            )
-            adj.det_meta.counter_records_per_thread[mapped_target] += meta.counter_records_per_thread.get(target, 1) - 1
-        det_args.append(mapped_target.helper_name if mapped_target is not None else target.helper_name)
-    for target in meta.scatter_targets:
-        mapped_label, mapped_attr_path = _deterministic_map_target(target, bound_args)
-        mapped_target = _deterministic_find_target(
-            adj.det_meta.scatter_targets, mapped_label, mapped_attr_path, adjoint=getattr(target, "adjoint", False)
-        )
-        if mapped_target is None and adj.det_registry is not None:
-            mapped_target = get_or_create_scatter_target(
-                adj.det_registry,
-                adj.det_meta,
-                mapped_label,
-                target.value_dtype,
-                target.value_ctype,
-                target.scalar_dtype,
-                target.reduce_op,
-                attr_path=mapped_attr_path,
-                adjoint=getattr(target, "adjoint", False),
-                use_forward=getattr(target, "use_forward", True),
-                use_backward=getattr(target, "use_backward", False),
-            )
-            adj.det_meta.scatter_records_per_thread[mapped_target] += meta.scatter_records_per_thread.get(target, 1) - 1
-        det_args.append(mapped_target.helper_name if mapped_target is not None else target.helper_name)
-    return det_args
-
-
-def _cinit_expr(dtype):
-    if hasattr(dtype, "cinit"):
-        return dtype.cinit(requires_grad=False)
-    return f"{Var.type_to_ctype(dtype)}{{}}"
 
 
 def codegen_kernel(kernel, device, options):
