+++ warp/_src/context.py
@@ -9087,517 +9087,6 @@ def _build_launch_bounds(dim: int | Sequence[int], kernel_dim: int) -> LaunchBou
     return _build_launch_bounds_from_tuple(dim, kernel_dim)
 
 
-class DeterministicLaunch(Launch):
-    """Recorded launch wrapper that replays through the deterministic launcher."""
-
-    def __init__(
-        self,
-        kernel,
-        device: Device,
-        det_meta,
-        fwd_args: Sequence[Any],
-        adj_args: Sequence[Any] | None = None,
-        hooks: KernelHooks | None = None,
-        params: Sequence[Any] | None = None,
-        bounds: launch_bounds_t | None = None,
-        max_blocks: int = 0,
-        block_dim: int = 256,
-        adjoint: bool = False,
-    ):
-        super().__init__(
-            kernel=kernel,
-            device=device,
-            hooks=hooks,
-            params=params,
-            params_addr=None,
-            bounds=bounds,
-            max_blocks=max_blocks,
-            block_dim=block_dim,
-            adjoint=adjoint,
-        )
-        self.det_meta = det_meta
-        self.fwd_args = list(fwd_args)
-        self.adj_args = list(adj_args) if adj_args is not None else []
-
-    def _is_deterministic_target_arg(self, index: int) -> bool:
-        if index < 0 or index >= len(self.kernel.adj.args):
-            return False
-
-        arg_label = self.kernel.adj.args[index].label
-        targets = (*self.det_meta.scatter_targets, *self.det_meta.counter_targets)
-        return any(target.array_var_label == arg_label for target in targets)
-
-    def set_param_at_index(self, index: int, value: Any, adjoint: bool = False):
-        super().set_param_at_index(index, value, adjoint)
-        if adjoint and index < len(self.adj_args):
-            self.adj_args[index] = value
-        elif not adjoint and index < len(self.fwd_args):
-            self.fwd_args[index] = value
-
-    def set_param_at_index_from_ctype(self, index: int, value: ctypes.Structure | int | float, adjoint: bool = False):
-        if self._is_deterministic_target_arg(index):
-            arg_label = self.kernel.adj.args[index].label
-            raise RuntimeError(
-                "Updating deterministic target argument "
-                f"'{arg_label}' from a raw ctypes value is not supported. "
-                "Use set_param_at_index() with a Warp array or struct object instead."
-            )
-
-        super().set_param_at_index_from_ctype(index, value, adjoint)
-        if adjoint and index < len(self.adj_args):
-            self.adj_args[index] = value
-        elif not adjoint and index < len(self.fwd_args):
-            self.fwd_args[index] = value
-
-    def launch(self, stream: Stream | None = None) -> None:
-        if stream is None:
-            stream = self.device.stream
-
-        _launch_deterministic(
-            self.kernel,
-            self.hooks,
-            self.params,
-            self.bounds,
-            self.device,
-            stream,
-            self.max_blocks,
-            self.block_dim,
-            self.det_meta,
-            self.fwd_args,
-            adj_args=self.adj_args,
-            adjoint=self.adjoint,
-            module_exec=self.module_exec,
-        )
-
-
-def _launch_deterministic(
-    kernel,
-    hooks,
-    user_params,
-    bounds,
-    device,
-    stream,
-    max_blocks,
-    block_dim,
-    det_meta,
-    fwd_args,
-    *,
-    adj_args=None,
-    adjoint=False,
-    module_exec=None,
-):
-    """Orchestrate a deterministic kernel launch with scatter-sort-reduce and/or two-pass execution.
-
-    This is called from launch() when deterministic mode is active and the
-    kernel has atomic operations that require deterministic treatment.
-    Scatter/reduce atomics use one kernel pass plus sort-reduce; consumed-return
-    counter atomics add a counting pass.
-    """
-    from warp._src.deterministic import (  # noqa: PLC0415
-        _det_dest_size_elements,
-        allocate_counter_buffers,
-        allocate_counter_state_buffers,
-        allocate_scatter_buffers,
-        get_counter_record_count,
-        get_scatter_record_count,
-        run_counter_scan,
-        run_counter_writeback,
-        run_sort_reduce,
-    )
-
-    launch_kind = "backward" if adjoint else "forward"
-
-    if hooks is None:
-        raise RuntimeError(
-            f"Failed to find {launch_kind} kernel '{kernel.key}' from module '{kernel.module.name}' for device '{device}'"
-        )
-
-    launch_hook = hooks.backward if adjoint else hooks.forward
-    shared_memory_bytes = hooks.backward_smem_bytes if adjoint else hooks.forward_smem_bytes
-
-    if launch_hook is None:
-        raise RuntimeError(
-            f"Failed to find {launch_kind} kernel '{kernel.key}' from module '{kernel.module.name}' for device '{device}'"
-        )
-
-    if runtime._apic_capture is not None:
-        raise RuntimeError(
-            "APIC serialization is not currently supported for deterministic CUDA kernels. "
-            "Capture with apic=False, or disable deterministic mode for kernels captured with apic=True."
-        )
-
-    adj_args = [] if adj_args is None else list(adj_args)
-
-    def resolve_det_target_array(target):
-        resolved = None
-        for j, arg in enumerate(kernel.adj.args):
-            if arg.label == target.array_var_label:
-                if getattr(target, "adjoint", False):
-                    resolved = adj_args[j] if j < len(adj_args) else None
-                else:
-                    resolved = fwd_args[j]
-                break
-
-        if resolved is None:
-            if not getattr(target, "adjoint", False):
-                return None
-
-            # Manual adjoint launches may rely on the forward array's embedded
-            # grad pointer instead of passing an explicit adjoint array. The
-            # deterministic post-pass needs the Python array object, so recover
-            # the same fallback here when possible.
-            primal = None
-            for j, arg in enumerate(kernel.adj.args):
-                if arg.label == target.array_var_label:
-                    primal = fwd_args[j]
-                    break
-            for attr in getattr(target, "attr_path", ()):
-                if primal is None or not hasattr(primal, attr):
-                    primal = None
-                    break
-                primal = getattr(primal, attr)
-            return getattr(primal, "grad", None)
-
-        for attr in getattr(target, "attr_path", ()):
-            if resolved is None:
-                return None
-            if not hasattr(resolved, attr):
-                raise RuntimeError(
-                    "Deterministic target "
-                    f"{target.target_label!r} could not resolve attribute {attr!r} "
-                    f"on {type(resolved).__name__}"
-                )
-            resolved = getattr(resolved, attr)
-
-        if getattr(target, "adjoint", False) and resolved is None:
-            primal = None
-            for j, arg in enumerate(kernel.adj.args):
-                if arg.label == target.array_var_label:
-                    primal = fwd_args[j]
-                    break
-            for attr in getattr(target, "attr_path", ()):
-                if primal is None or not hasattr(primal, attr):
-                    primal = None
-                    break
-                primal = getattr(primal, attr)
-            resolved = getattr(primal, "grad", None)
-
-        return resolved
-
-    def target_has_resolvable_destination(target):
-        array = resolve_det_target_array(target)
-        if array is None:
-            return False
-        if not hasattr(array, "ptr"):
-            raise RuntimeError(
-                f"Deterministic target {target.target_label!r} resolved to non-array object ({type(array).__name__})"
-            )
-        return getattr(array, "ptr", None) is not None and getattr(array, "size", 0) > 0
-
-    active_scatter_targets = []
-    for target in det_meta.scatter_targets:
-        if not (target.use_backward if adjoint else target.use_forward):
-            continue
-        if adjoint and not target_has_resolvable_destination(target):
-            continue
-        active_scatter_targets.append(target)
-
-    active_counter_targets = []
-    for target in det_meta.counter_targets:
-        if not (target.use_backward if adjoint else target.use_forward):
-            continue
-        if adjoint and not target_has_resolvable_destination(target):
-            continue
-        active_counter_targets.append(target)
-
-    dim_size = bounds.size
-    if active_scatter_targets and dim_size > (1 << 32):
-        raise RuntimeError(
-            "Deterministic scatter atomics support launch sizes up to 2^32 threads because sort keys pack the "
-            "linear thread index into 32 bits."
-        )
-    if active_counter_targets and dim_size > ((1 << 31) - 1):
-        raise RuntimeError(
-            "Deterministic consumed-return counter atomics support launch sizes up to 2^31 - 1 threads because "
-            "counter prefix buffers store per-thread contributions as int32 values."
-        )
-
-    determinism_mode = det_meta.determinism_mode
-    max_records = det_meta.max_records
-    det_debug = int(warp.config.deterministic_debug)
-    stream_is_capturing = len(runtime.captures) > 0 and runtime.core.wp_cuda_stream_is_capturing(stream.cuda_stream)
-    capture_graph = None
-    if stream_is_capturing:
-        capture_id = runtime.core.wp_cuda_stream_get_capture_id(stream.cuda_stream)
-        capture_graph = runtime.captures.get(capture_id)
-
-    # Allocate buffers.
-    scatter_bufs = (
-        allocate_scatter_buffers(
-            active_scatter_targets,
-            det_meta,
-            dim_size,
-            device,
-            max_records=max_records,
-            initialize_unused=stream_is_capturing,
-        )
-        if active_scatter_targets
-        else []
-    )
-    counter_bufs = (
-        allocate_counter_buffers(
-            active_counter_targets,
-            det_meta,
-            dim_size,
-            device,
-            max_records=max_records,
-            initialize_unused=stream_is_capturing,
-        )
-        if active_counter_targets
-        else []
-    )
-    scatter_bufs_by_target = dict(zip(active_scatter_targets, scatter_bufs, strict=True))
-    counter_bufs_by_target = dict(zip(active_counter_targets, counter_bufs, strict=True))
-    overflow_buf = warp.zeros(shape=(1,), dtype=warp.int32, device=device) if scatter_bufs or counter_bufs else None
-    sort_reduce_workspaces = []
-    counter_scan_workspaces = []
-    counter_state_buffers = []
-    counter_arr_by_target = {}
-
-    # Build the extra deterministic parameters (must match codegen_kernel order).
-    def build_det_params(phase, use_scatter):
-        det_params = [
-            ctypes.c_int(phase),
-            ctypes.c_int(det_debug),
-            ctypes.c_void_p(overflow_buf.ptr if overflow_buf is not None else 0),
-        ]
-        for ct in det_meta.counter_targets:
-            counter_buf = counter_bufs_by_target.get(ct)
-            counter_arr = counter_arr_by_target.get(ct)
-            target_ptr = counter_arr.ptr if counter_arr is not None else 0
-            target_size = _det_dest_size_elements(counter_arr) if counter_arr is not None else 0
-            if counter_buf is None:
-                det_params.append(det_counter_buf_t(0, 0, 0, 0, 0, 0, 0, 0, target_ptr, target_size))
-                continue
-
-            keys, values, prefixes, record_slots, cursors, count, capacity, records_per_thread = counter_buf
-            det_params.append(
-                det_counter_buf_t(
-                    keys.ptr,
-                    values.ptr,
-                    prefixes.ptr,
-                    record_slots.ptr,
-                    cursors.ptr,
-                    count.ptr,
-                    capacity,
-                    records_per_thread,
-                    target_ptr,
-                    target_size,
-                )
-            )
-        for st in det_meta.scatter_targets:
-            scatter_buf = scatter_bufs_by_target.get(st)
-            if use_scatter and scatter_buf is not None:
-                keys, values, counter, capacity = scatter_buf
-                det_params.append(det_scatter_buf_t(keys.ptr, values.ptr, counter.ptr, capacity))
-            else:
-                # Null scatter buffers (phase 0 doesn't scatter).
-                det_params.append(det_scatter_buf_t(0, 0, 0, 0))
-        return det_params
-
-    def do_cuda_launch(hook, params_list):
-        _cuda_launch_kernel(
-            device,
-            module_exec,
-            hook,
-            bounds.size,
-            max_blocks,
-            block_dim,
-            shared_memory_bytes,
-            _build_cuda_kernel_params(params_list),
-            stream,
-        )
-
-    if active_counter_targets:
-        # === Two-pass execution ===
-
-        counter_arrays = []
-        for ct in active_counter_targets:
-            counter_arr = resolve_det_target_array(ct)
-            if counter_arr is None:
-                counter_arrays.append(None)
-                continue
-            if not hasattr(counter_arr, "ptr"):
-                raise RuntimeError(
-                    "Deterministic counter target "
-                    f"{ct.target_label!r} resolved to non-array object "
-                    f"({type(counter_arr).__name__})"
-                )
-            if getattr(counter_arr, "ptr", None) is None or getattr(counter_arr, "size", 0) <= 0:
-                counter_arrays.append(None)
-                continue
-            counter_arrays.append(counter_arr)
-
-        for ct, counter_arr in zip(active_counter_targets, counter_arrays, strict=True):
-            counter_arr_by_target[ct] = counter_arr
-
-        with warp.ScopedStream(stream, sync_enter=False):
-            for counter_buf in counter_bufs:
-                keys, values, _prefixes, record_slots, cursors, count, _capacity, _records_per_thread = counter_buf
-                count.zero_()
-                cursors.zero_()
-                record_slots.fill_(-1)
-                if stream_is_capturing:
-                    keys.fill_(-1)
-                    values.zero_()
-
-        # Phase 0: counting pass (side effects suppressed, scatter disabled).
-        det_params_p0 = build_det_params(phase=0, use_scatter=False)
-        # Append det params after all user args (matches codegen_kernel order:
-        # dim, user_args..., det_params...).
-        params_p0 = [*user_params, *det_params_p0]
-        do_cuda_launch(launch_hook, params_p0)
-
-        with warp.ScopedStream(stream, sync_enter=False):
-            counter_state_buffers = allocate_counter_state_buffers(runtime, counter_arrays, device, stream)
-
-        # Sort counter records by destination and deterministic record order,
-        # then compute the exclusive prefix each atomic should return.
-        counter_record_counts = []
-        for i, counter_arr in enumerate(counter_arrays):
-            ct = active_counter_targets[i]
-            if counter_arr is None:
-                if not stream_is_capturing:
-                    _keys, _values, _prefixes, _record_slots, _cursors, count, _capacity, _records_per_thread = (
-                        counter_bufs[i]
-                    )
-                    emitted_count = int(count.numpy()[0])
-                    if emitted_count != 0:
-                        raise RuntimeError(
-                            f"Deterministic counter target {ct.target_label!r} resolved to None "
-                            f"but emitted {emitted_count} records."
-                        )
-                counter_record_counts.append(0)
-                continue
-            if stream_is_capturing:
-                record_count = get_counter_record_count(counter_bufs[i], stream_is_capturing=True)
-            else:
-                _keys, _values, _prefixes, _record_slots, _cursors, count, capacity, _records_per_thread = counter_bufs[
-                    i
-                ]
-                emitted_count = int(count.numpy()[0])
-                if emitted_count < 0 or emitted_count > capacity:
-                    raise RuntimeError(
-                        f"Deterministic counter buffer overflow in kernel '{kernel.key}'. "
-                        "Increase 'deterministic_max_records' or reduce the per-thread atomic count."
-                    )
-                record_count = emitted_count
-            counter_record_counts.append(record_count)
-
-        if overflow_buf is not None and not stream_is_capturing and int(overflow_buf.numpy()[0]) != 0:
-            raise RuntimeError(
-                f"Deterministic counter buffer overflow in kernel '{kernel.key}'. "
-                "Increase 'deterministic_max_records' or reduce the per-thread atomic count."
-            )
-
-        with warp.ScopedStream(stream, sync_enter=False):
-            counter_scan_workspaces = run_counter_scan(
-                runtime,
-                counter_bufs,
-                counter_arrays,
-                counter_state_buffers,
-                device,
-                record_counts=counter_record_counts,
-            )
-            for counter_buf in counter_bufs:
-                _keys, _values, _prefixes, _record_slots, cursors, _count, _capacity, _records_per_thread = counter_buf
-                cursors.zero_()
-
-        # Phase 1: execution pass with deterministic slots.
-        det_params_p1 = build_det_params(phase=1, use_scatter=True)
-        params_p1 = [*user_params, *det_params_p1]
-        do_cuda_launch(launch_hook, params_p1)
-
-        with warp.ScopedStream(stream, sync_enter=False):
-            run_counter_writeback(
-                runtime,
-                counter_bufs,
-                counter_arrays,
-                counter_state_buffers,
-                record_counts=counter_record_counts,
-            )
-
-    else:
-        # === Single-pass (scatter only, or inactive deterministic params only) ===
-        det_params = build_det_params(phase=1, use_scatter=True)
-        params_all = [*user_params, *det_params]
-        do_cuda_launch(launch_hook, params_all)
-
-    # Post-kernel: reduce scatter records in a fixed order.
-    if active_scatter_targets:
-        # Identify the destination arrays from fwd_args.
-        scatter_targets = []
-        scatter_buffers = []
-        dest_arrays = []
-        record_counts = []
-        for scatter_target, scatter_buffer in zip(active_scatter_targets, scatter_bufs, strict=True):
-            dest_array = resolve_det_target_array(scatter_target)
-            if dest_array is None:
-                continue
-            if not hasattr(dest_array, "ptr"):
-                raise RuntimeError(
-                    "Deterministic scatter target "
-                    f"{scatter_target.target_label!r} resolved to non-array object "
-                    f"({type(dest_array).__name__})"
-                )
-            if getattr(dest_array, "ptr", None) is None or getattr(dest_array, "size", 0) <= 0:
-                continue
-            scatter_targets.append(scatter_target)
-            scatter_buffers.append(scatter_buffer)
-            dest_arrays.append(dest_array)
-            record_counts.append(get_scatter_record_count(scatter_buffer, stream_is_capturing=stream_is_capturing))
-
-        with warp.ScopedStream(stream, sync_enter=False):
-            sort_reduce_workspaces = run_sort_reduce(
-                runtime,
-                scatter_targets,
-                scatter_buffers,
-                dest_arrays,
-                device,
-                determinism_mode,
-                record_counts=record_counts,
-            )
-
-    if capture_graph is not None:
-        # Captured graphs replay after this function returns, so keep temporary
-        # deterministic buffers alive for the lifetime of the graph object.
-        capture_graph._deterministic_buffer_refs.extend(
-            buffer
-            for buffer in (
-                *scatter_bufs,
-                *counter_bufs,
-                overflow_buf,
-                *sort_reduce_workspaces,
-                *counter_scan_workspaces,
-                *(array for state_buffer in counter_state_buffers for array in state_buffer[:2] if array is not None),
-            )
-            if buffer is not None
-        )
-
-    try:
-        runtime.verify_cuda_device(device)
-    except Exception:
-        warp._src.utils.warn(f"Error in deterministic kernel launch: {kernel.key} on device {device}")
-        raise
-
-    if overflow_buf is not None and not stream_is_capturing and int(overflow_buf.numpy()[0]) != 0:
-        overflow_kind = "scatter" if det_meta.has_scatter else "counter"
-        raise RuntimeError(
-            f"Deterministic {overflow_kind} buffer overflow in kernel '{kernel.key}'. "
-            "Increase 'deterministic_max_records' or reduce the per-thread atomic count."
-        )
-
-
 def launch(
     kernel,
     dim: int | Sequence[int],
@@ -9720,6 +9209,8 @@ def launch(
         # tape adjoints can reduce gradient atomics in a fixed order.
         det_meta = getattr(kernel.adj, "det_meta", None)
         if det_meta is not None and det_meta.needs_deterministic and device.is_cuda:
+            from warp._src.deterministic import create_deterministic_launch, launch_deterministic  # noqa: PLC0415
+
             launch_hook = hooks.backward if adjoint else hooks.forward
             launch_kind = "backward" if adjoint else "forward"
             if launch_hook is None:
@@ -9729,7 +9220,7 @@ def launch(
             if stream is None:
                 stream = device.stream
             if record_cmd:
-                return DeterministicLaunch(
+                return create_deterministic_launch(
                     kernel=kernel,
                     device=device,
                     det_meta=det_meta,
@@ -9742,7 +9233,7 @@ def launch(
                     block_dim=block_dim,
                     adjoint=adjoint,
                 )
-            _launch_deterministic(
+            launch_deterministic(
                 kernel,
                 hooks,
                 params,
